backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai>=1.30.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
